shopFusion/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==3.2.1
django-paypal==2.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
twilio==9.0.2
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
